refactor(workflows): fold search registration into search-sync (C3c)
Final phase of the workflows/index.ts split. Moves the inline
registerFileBackedContentType block out of activate() into search-sync as
registerWorkflowSearch(ctx) — definitions + instances file patterns, the
onUnlink shadow-fallback, the full reindex generator, and verifyExists, all
beside the search-doc builders they already use.

activate() now just calls registerWorkflowSearch(ctx). index.ts sheds the
fs / yaml / path-getManagedDefinition / loadDefinition / WorkflowDefinition /
WorkflowInstance imports the block solely used, and drops from 288 to 169
lines — pure activate-orchestration + lifecycle.

This completes the workflows/index.ts decomposition: 2,146 -> 169 (92%).

Pure behavior-preserving relocation — the content-type config is
byte-identical.

Verified: typecheck + eslint + workflows suite (420/0) + full suite (5124/0)
+ binary build + isolated-home boot smoke (bakin_workflows content type
registers; /definitions + /instances + /gates/pending + /search → 200).
