+++ plugins/workflows/index.ts
@@ -3,13 +3,11 @@
  * Enforces step-by-step agent execution with gated delivery,
  * parallel steps, human gates, and output validation.
  */
-import { existsSync, readdirSync, readFileSync } from 'fs'
 import { join, dirname } from 'path'
 import { fileURLToPath } from 'url'
-import yaml from 'js-yaml'
 import type { BakinPlugin, PluginContext } from '@bakin/core/plugin-types'
 import { definePlugin } from '@bakin/core/routing'
-import { listDefinitions, loadDefinition } from './lib/parser'
+import { listDefinitions } from './lib/parser'
 import { loadDefaultWorkflows } from './lib/load-defaults'
 import {
   checkWorkflowDefinitions,
@@ -18,10 +16,9 @@ import {
   workflowSkillDriftRepair,
   staleWorkflowInstancesRepair,
 } from './lib/health-checks'
-import { getManagedDefinition } from './lib/source-registry'
 import { listInstances, reconcilePendingApprovalTaskColumns } from './lib/runtime'
 import { setWorkflowPluginContext } from './lib/plugin-context'
-import { instanceToSearchDoc, definitionToSearchDoc } from './lib/search-sync'
+import { registerWorkflowSearch } from './lib/search-sync'
 import { definitionRoutes } from './lib/routes/definitions'
 import { instanceRoutes } from './lib/routes/instances'
 import { gateRoutes } from './lib/routes/gates'
@@ -37,7 +34,6 @@ import {
   setNotificationRuntime,
   type GateNotificationSettings,
 } from './lib/notifications'
-import type { WorkflowDefinition, WorkflowInstance } from './types'
 
 const log = createLogger('workflows')
 let unsubscribeApprovalResponses: (() => void) | null = null
@@ -69,122 +65,7 @@ const workflowsPlugin: BakinPlugin = definePlugin({
     setWorkflowPluginContext(ctx)
 
     // ─── Search Content Type Registration ─────────────────────────────
-
-    ctx.search.registerFileBackedContentType({
-      table: 'workflows',
-      schema: {
-        name: { type: 'text' },
-        description: { type: 'text' },
-        type: { type: 'keyword' },
-        status: { type: 'keyword' },
-        task_id: { type: 'keyword' },
-        steps: { type: 'text' },
-        updated_at: { type: 'datetime' },
-      },
-      searchableFields: ['name', 'description', 'steps'],
-      rerankField: 'description',
-      embeddingTemplate: '{{name}} {{description}} {{steps}}',
-      facets: ['type', 'status'],
-      filePatterns: [
-        {
-          pattern: 'workflows/definitions/*.{yaml,yml}',
-          fileToId: (rel) => {
-            const name = rel.replace(/^workflows\/definitions\//, '').replace(/\.(yaml|yml)$/, '')
-            return `def:${name}`
-          },
-          fileToDoc: async (rel) => {
-            const name = rel.replace(/^workflows\/definitions\//, '').replace(/\.(yaml|yml)$/, '')
-            const def = loadDefinition(name)
-            return def ? definitionToSearchDoc(name, def, def.source) : null
-          },
-        },
-        {
-          pattern: 'workflows/instances/*.json',
-          fileToId: (rel) => {
-            const taskId = rel.replace(/^workflows\/instances\//, '').replace(/\.json$/, '')
-            return `inst:${taskId}`
-          },
-          fileToDoc: async (rel, content) => {
-            try {
-              const data = JSON.parse(content) as WorkflowInstance
-              return instanceToSearchDoc(data)
-            } catch {
-              return null
-            }
-          },
-        },
-      ],
-      preserveVirtualDocuments: true,
-      onUnlink: async (rel) => {
-        if (rel.startsWith('workflows/definitions/')) {
-          const name = rel.replace(/^workflows\/definitions\//, '').replace(/\.(yaml|yml)$/, '')
-          const defsDir = join(getContentDir(), 'workflows', 'definitions')
-          const alternateUserPath = rel.endsWith('.yaml')
-            ? join(defsDir, `${name}.yml`)
-            : join(defsDir, `${name}.yaml`)
-
-          if (existsSync(alternateUserPath)) {
-            const alternateDefinition = yaml.load(readFileSync(alternateUserPath, 'utf-8')) as WorkflowDefinition
-            await ctx.search.index(
-              `def:${name}`,
-              definitionToSearchDoc(name, { ...alternateDefinition, source: 'user' }, 'user'),
-            )
-            return
-          }
-
-          const fallbackEntry = getManagedDefinition(name)
-          if (fallbackEntry) {
-            await ctx.search.index(
-              `def:${name}`,
-              definitionToSearchDoc(
-                name,
-                { ...fallbackEntry.definition, source: fallbackEntry.source },
-                fallbackEntry.source,
-              ),
-            )
-          } else {
-            await ctx.search.remove(`def:${name}`)
-          }
-          return
-        }
-        if (rel.startsWith('workflows/instances/')) {
-          const taskId = rel.replace(/^workflows\/instances\//, '').replace(/\.json$/, '')
-          await ctx.search.remove(`inst:${taskId}`)
-        }
-      },
-      reindex: async function* () {
-        const contentDir = getContentDir()
-
-        // Yield effective definitions from every source: plugin defaults,
-        // agent-package definitions, and user YAML shadows.
-        for (const entry of listDefinitions(contentDir)) {
-          yield { key: `def:${entry.name}`, doc: definitionToSearchDoc(entry.name, entry.definition, entry.source) }
-        }
-
-        // Yield instances
-        const instancesDir = join(contentDir, 'workflows', 'instances')
-        if (existsSync(instancesDir)) {
-          for (const file of readdirSync(instancesDir).filter(f => f.endsWith('.json'))) {
-            try {
-              const data = JSON.parse(readFileSync(join(instancesDir, file), 'utf-8')) as WorkflowInstance
-              yield { key: `inst:${data.taskId}`, doc: instanceToSearchDoc(data) }
-            } catch { /* skip corrupt instances */ }
-          }
-        }
-      },
-      verifyExists: async (key: string) => {
-        const contentDir = getContentDir()
-        if (key.startsWith('def:')) {
-          const name = key.slice(4)
-          return loadDefinition(name, contentDir) !== null
-        }
-        if (key.startsWith('inst:')) {
-          const taskId = key.slice(5)
-          return existsSync(join(contentDir, 'workflows', 'instances', `${taskId}.json`))
-        }
-        return false
-      },
-    })
+    registerWorkflowSearch(ctx)
 
     // ─── Plugin-shipped workflow defaults ─────────────────────────────
     // Load every YAML in defaults/workflows/ and register through
